Blocks introduce implicit label indices.
